+++ crates/notify-kit/src/lib.rs
@@ -1,5 +1,7 @@
 #![forbid(unsafe_code)]
 
+pub mod builtin;
+pub mod core;
 pub mod env;
 mod error;
 mod event;
@@ -7,34 +9,30 @@ mod hub;
 mod log;
 mod sinks;
 
-pub use crate::error::{Error, ErrorKind, SinkFailure};
-pub type Result<T> = std::result::Result<T, Error>;
-
-pub use crate::event::{Event, Severity};
-pub use crate::hub::{Hub, HubConfig, HubLimits, TryNotifyError};
-pub use crate::sinks::Sink;
 #[cfg(any(not(feature = "selective-sinks"), feature = "bark"))]
-pub use crate::sinks::{BarkConfig, BarkSink};
+pub use crate::builtin::{BarkConfig, BarkSink};
 #[cfg(any(not(feature = "selective-sinks"), feature = "dingtalk"))]
-pub use crate::sinks::{DingTalkWebhookConfig, DingTalkWebhookSink};
+pub use crate::builtin::{DingTalkWebhookConfig, DingTalkWebhookSink};
 #[cfg(any(not(feature = "selective-sinks"), feature = "discord"))]
-pub use crate::sinks::{DiscordWebhookConfig, DiscordWebhookSink};
+pub use crate::builtin::{DiscordWebhookConfig, DiscordWebhookSink};
 #[cfg(any(not(feature = "selective-sinks"), feature = "feishu"))]
-pub use crate::sinks::{FeishuWebhookConfig, FeishuWebhookSink};
+pub use crate::builtin::{FeishuWebhookConfig, FeishuWebhookSink};
 #[cfg(any(not(feature = "selective-sinks"), feature = "generic-webhook"))]
-pub use crate::sinks::{GenericWebhookConfig, GenericWebhookSink};
+pub use crate::builtin::{GenericWebhookConfig, GenericWebhookSink};
 #[cfg(any(not(feature = "selective-sinks"), feature = "github"))]
-pub use crate::sinks::{GitHubCommentConfig, GitHubCommentSink};
+pub use crate::builtin::{GitHubCommentConfig, GitHubCommentSink};
 #[cfg(any(not(feature = "selective-sinks"), feature = "pushplus"))]
-pub use crate::sinks::{PushPlusConfig, PushPlusSink};
+pub use crate::builtin::{PushPlusConfig, PushPlusSink};
 #[cfg(any(not(feature = "selective-sinks"), feature = "serverchan"))]
-pub use crate::sinks::{ServerChanConfig, ServerChanSink};
+pub use crate::builtin::{ServerChanConfig, ServerChanSink};
 #[cfg(any(not(feature = "selective-sinks"), feature = "slack"))]
-pub use crate::sinks::{SlackWebhookConfig, SlackWebhookSink};
+pub use crate::builtin::{SlackWebhookConfig, SlackWebhookSink};
 #[cfg(any(not(feature = "selective-sinks"), feature = "sound"))]
-pub use crate::sinks::{SoundConfig, SoundSink};
+pub use crate::builtin::{SoundConfig, SoundSink};
 #[cfg(any(not(feature = "selective-sinks"), feature = "telegram"))]
-pub use crate::sinks::{TelegramBotConfig, TelegramBotSink};
+pub use crate::builtin::{TelegramBotConfig, TelegramBotSink};
 #[cfg(any(not(feature = "selective-sinks"), feature = "wecom"))]
-pub use crate::sinks::{WeComWebhookConfig, WeComWebhookSink};
+pub use crate::builtin::{WeComWebhookConfig, WeComWebhookSink};
+pub use crate::core::{Error, ErrorKind, Event, Hub, HubConfig, HubLimits, Result, Severity};
+pub use crate::core::{Sink, SinkFailure, TryNotifyError};
 pub use secret_kit::SecretString;